add ParameterPrefix,QuotePrefix,QuoteSuffix test
